feat(workspace): port chunker to dasclaw_workspace_cap (#630)
Verbatim move of the document-chunking helper from ironclaw to the
shared dasclaw_workspace_cap crate. This is slice 1/N of ADR-152 §3
F3.4 (the lightest, zero-dependency module first).

Moved (verbatim, ADR-129 §1.3):
- ChunkConfig (with Default + with_chunk_size + with_overlap)
- chunk_document
- 6 unit tests

Compatibility:
- ironclaw side becomes a 5-line re-export shim
- 'pub use chunker::{ChunkConfig, chunk_document}' in workspace/mod.rs
  keeps the public path 'ironclaw::workspace::{ChunkConfig, chunk_document}'
- No call-site changes anywhere
- No new dependencies (chunker uses only std)

Not in this PR (deferred to later slices):
- document / embedding_cache / embeddings / hygiene / layer / privacy /
  repository / search / mod — each has heavier deps and will land
  separately

Cross-cuts: 类A (no new .ironclaw / IRONCLAW_BASE_DIR literals)

Refs #630
